geotags-api/requirements.txt
Flask==0.10.1
Flask-Cors==2.1.2
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-SeaSurf==0.2.1
Flask-SQLAlchemy==2.1
Flask-User==0.6.8
Flask-WTF==0.12
GeoAlchemy2==0.2.6
geojson==1.3.2
gunicorn==19.4.5
ModestMaps==1.4.6
Shapely==1.5.13
TileStache==1.50.1
psycopg2==2.6.1
Flask-Babel==0.9
pylru==1.0.9
